refactor: msw 폴더 구조 co-location으로 리팩터링 #190
