+++ frontend/src/common/mock/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
 
-import { handlers } from './handlers/handlers';
+import { handlers } from './handlers';
 
 export const worker = setupWorker(...handlers);
