+++ frontend/src/common/mock/handlers.ts
@@ -1,0 +1,3 @@
+import { mentoringHandler } from './mentoring/handlers';
+
+export const handlers = [...mentoringHandler];
