--- frontend/src/common/mock/handlers/handlers.ts
@@ -1,3 +1,0 @@
-import { mentoringHandler } from './mentoring';
-
-export const handlers = [...mentoringHandler];
