+++ frontend/src/common/mock/server.ts
@@ -1,5 +1,5 @@
 import { setupServer } from 'msw/node';
 
-import { handlers } from './handlers/handlers';
+import { handlers } from './handlers';
 
 export const server = setupServer(...handlers);
